# docs/swagger.yaml
definitions:
  errorcode.Error:
    type: object
...
  requests.HealthCheckDetail:
    properties:
      command:
        description: Type=Command 时使用
        type: string
      failure_threshold:
        type: integer
      http_header:
        items:
          $ref: '#/definitions/requests.HttpHeader'
        type: array
      initial_delay_seconds:
        type: integer
      path:
        type: string
      period_seconds:
        type: integer
      port:
        type: integer
      protocol:
        type: string
      success_threshold:
        type: integer
      timeout_seconds:
        type: integer
      type:
        description: HTTP|TCP|Command
        type: string
    type: object
...
  requests.KubeApplyYamlRequest:
    properties:
      name:
        type: string
      namespace:
        type: string
      yaml:
        type: string
    type: object
...
  requests.KubeStatefulSetCreateRequest:
    properties:
      container_command:
        description: 容器启动命令
        type: string
      container_command_args:
        description: 容器启动参数
        type: string
      container_image:
        description: 主容器镜像
        type: string
      cpu_requirement:
        description: CPU 需求（可选）
        type: string
      description:
        description: 描述信息（可选）
        type: string
      image_pull_secret:
        description: 镜像/启动相关（可选）
        type: string
      is_create_service:
        description: Service 相关（StatefulSet 必须绑定一个服务名，通常为 Headless Service）
        type: boolean
      is_liveness_enable:
        description: 是否启用 Liveness
        type: boolean
      is_readiness_enable:
        description: 是否启用 Readiness
        type: boolean
      labels:
        description: 标签
        items:
          $ref: '#/definitions/requests.Label'
        type: array
      liveness_probe:
        allOf:
        - $ref: '#/definitions/requests.HealthCheckDetail'
        description: Liveness 配置
      memory_requirement:
        description: 资源/标签/环境变量/端口
        type: string
      name:
        description: 基本信息
        type: string
      namespace:
        description: 命名空间
        type: string
      port_mappings:
        description: 端口映射
        items:
          $ref: '#/definitions/requests.PortMapping'
        type: array
      readiness_probe:
        allOf:
        - $ref: '#/definitions/requests.HealthCheckDetail'
        description: Readiness 配置
      replicas:
        description: 副本数
        type: integer
      run_as_privileged:
        description: 探针/特权
        type: boolean
      service_name:
        description: 必填：绑定的 Service 名（Headless 推荐）
        type: string
      service_type:
        description: Service 类型（ClusterIP/NodePort/...；Headless 时忽略）
        type: string
      update_partition:
        description: 滚动更新策略（可选）
        type: integer
      variables:
        description: 环境变量
        items:
          $ref: '#/definitions/requests.EnvironmentVariable'
        type: array
      volume_claim_templates:
        description: 每个 Pod 的独立 PVC（StatefulSet 常用）
        items:
          $ref: '#/definitions/requests.VolumeClaimTemplate'
        type: array
    type: object
  requests.KubeStatefulSetRestartRequest:
    properties:
      name:
        type: string
      namespace:
        type: string
    type: object
  requests.KubeStatefulSetRollbackRequest:
    properties:
      name:
        type: string
      namespace:
        type: string
      revision_name:
        description: 指定回滚到的历史版本（ControllerRevision）
        type: string
    type: object
  requests.KubeStatefulSetScaleRequest:
    properties:
      name:
        type: string
      namespace:
        type: string
      scale_num:
        description: 副本数量
        type: integer
    type: object
  requests.KubeStatefulSetUpdateImageRequest:
    properties:
      container:
        description: 目标容器名称
        type: string
      image:
        description: 新镜像地址，例如 mysql:8.0
        type: string
      name:
        type: string
      namespace:
        type: string
    type: object
...
  requests.Label:
    properties:
      key:
        type: string
      value:
        type: string
    type: object
...
  requests.YamlCreateRequest:
    properties:
      yaml:
        description: YAML 内容
        type: string
    required:
    - yaml
    type: object
  response.Response:
    type: object
info:
  contact: {}
  description: 基于Gin+Vue开发的K8s管理平台
  termsOfService: https://gitee.com/jay-kim/k8s_operation
  title: K8s管理平台
  version: "1.0"
paths:
...
  /api/v1/k8s/statefulset/apply_yaml:
    put:
      consumes:
      - application/json
      description: 应用修改后的 YAML 配置到 StatefulSet
      parameters:
      - description: YAML内容
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/requests.KubeApplyYamlRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errorcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errorcode.Error'
      summary: 应用 StatefulSet YAML 配置
      tags:
      - K8s StatefulSet 管理
  /api/v1/k8s/statefulset/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/requests.KubeStatefulSetCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorcode.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorcode.Error'
      summary: 创建 StatefulSet（可选创建 Service）
      tags:
      - K8s StatefulSet 管理
  /api/v1/k8s/statefulset/create-from-yaml:
    post:
      consumes:
      - application/json
      parameters:
      - description: YAML 创建参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/requests.YamlCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorcode.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorcode.Error'
      summary: 从 YAML 创建 StatefulSet
      tags:
      - K8s StatefulSet 管理
  /api/v1/k8s/statefulset/delete:
    delete:
      consumes:
      - application/json
      description: 前台级联删除 StatefulSet（先删 Pod/ControllerRevision，再删 StatefulSet 本体）；成功返回命名空间、名称与状态信息
      parameters:
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      - description: StatefulSet 名称
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '示例: {\"namespace\":\"default\",\"name\":\"web\",\"status\":\"StatefulSet
            删除成功\"}'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/errorcode.Error'
        "404":
          description: StatefulSet 未找到
          schema:
            $ref: '#/definitions/errorcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errorcode.Error'
      security:
      - BearerAuth: []
      summary: 删除 StatefulSet
      tags:
      - K8s StatefulSet 管理
  /api/v1/k8s/statefulset/delete_svc:
    delete:
      consumes:
      - application/json
      description: 根据命名空间与名称删除 Service；成功仅表示删除请求已受理。
      parameters:
      - description: 删除参数（namespace、name）
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/errorcode.Error'
        "404":
          description: Service 未找到
          schema:
            $ref: '#/definitions/errorcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errorcode.Error'
      summary: 删除 Service
      tags:
      - K8s StatefulSet 管理
  /api/v1/k8s/statefulset/detail:
    get:
      consumes:
      - application/json
      description: 根据命名空间和名称获取单个 StatefulSet 的详细信息
      parameters:
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      - description: StatefulSet 名称
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取详情成功
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorcode.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorcode.Error'
      summary: 获取 StatefulSet 详情
      tags:
      - K8s StatefulSet 管理
  /api/v1/k8s/statefulset/events:
    post:
      parameters:
      - description: 命名空间
        in: query
        name: namespace
        type: string
      - description: 资源类型
        in: query
        name: kind
        type: string
      - description: 资源名称
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 事件列表加载完成
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorcode.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorcode.Error'
      summary: 获取事件列表（支持全局或按对象筛选）
      tags:
      - K8s StatefulSet 管理
  /api/v1/k8s/statefulset/history:
    get:
      parameters:
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      - description: StatefulSet 名称
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorcode.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorcode.Error'
      summary: 获取 StatefulSet 历史版本（ControllerRevision 列表）
      tags:
      - K8s StatefulSet 管理
  /api/v1/k8s/statefulset/list:
    get:
      consumes:
      - application/json
      description: 分页、模糊查询 StatefulSet 列表
      parameters:
      - description: StatefulSet 名称关键字（模糊匹配）
        in: query
        name: name
        type: string
      - description: 页码（默认 1）
        in: query
        name: page
        type: integer
      - description: 每页数量（默认 10）
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorcode.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorcode.Error'
      summary: 获取 StatefulSet 列表
      tags:
      - K8s StatefulSet 管理
  /api/v1/k8s/statefulset/restart:
    post:
      consumes:
      - application/json
      description: 在 .spec.template.metadata.annotations 写入 `kubectl.kubernetes.io/restartedAt`
        时间戳，从而触发 StatefulSet 的滚动更新；等价于 `kubectl rollout restart sts <name>`。
      parameters:
      - description: 重启参数（namespace、name）
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/requests.KubeStatefulSetRestartRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 重启成功，返回命名空间、名称与触发时间等信息
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/errorcode.Error'
        "404":
          description: StatefulSet 未找到
          schema:
            $ref: '#/definitions/errorcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errorcode.Error'
      summary: 重启 StatefulSet（触发滚动重启）
      tags:
      - K8s StatefulSet 管理
  /api/v1/k8s/statefulset/rollback:
    post:
      consumes:
      - application/json
      description: 将 StatefulSet 回滚到指定的历史版本（ControllerRevision）
      parameters:
      - description: 回滚参数（namespace、name、revision_name）
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.KubeStatefulSetRollbackRequest'
      produces:
      - application/json
      responses:
        "200":
          description: StatefulSet 回滚成功
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorcode.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorcode.Error'
      summary: 回滚 StatefulSet
      tags:
      - K8s StatefulSet 管理
  /api/v1/k8s/statefulset/scale:
    put:
      consumes:
      - application/json
      description: 通过 Patch 局部更新 .spec.replicas，K8s 将按策略有序创建/删除 Pod
      parameters:
      - description: 扩缩容参数（namespace、name、scale_num）
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/requests.KubeStatefulSetScaleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 扩缩容成功，返回修改前后及当前副本信息
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorcode.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorcode.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorcode.Error'
      summary: 扩缩容 StatefulSet（修改副本数）
      tags:
      - K8s StatefulSet 管理
  /api/v1/k8s/statefulset/sts_pods:
    get:
      parameters:
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      - description: StatefulSet 名称
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorcode.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorcode.Error'
      summary: 获取 StatefulSet 对应的 Pod 列表
      tags:
      - K8s StatefulSet 管理
  /api/v1/k8s/statefulset/update_image:
    put:
      consumes:
      - application/json
      description: 仅修改 .spec.template.spec.containers[*].image，不影响其它字段；根据 UpdateStrategy
        触发滚动更新
      parameters:
      - description: 更新镜像参数（namespace、name、container、image）
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/requests.KubeStatefulSetUpdateImageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功，返回资源版本与副本进度
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorcode.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorcode.Error'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/errorcode.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorcode.Error'
      summary: 更新 StatefulSet 容器镜像（Patch 局部更新）
      tags:
      - K8s StatefulSet 管理
  /api/v1/k8s/statefulset/yaml:
    get:
      description: 获取指定 StatefulSet 的 YAML 格式配置
      parameters:
      - description: 命名空间
        in: query
        name: namespace
        required: true
        type: string
      - description: StatefulSet 名称
        in: query
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errorcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errorcode.Error'
      summary: 获取 StatefulSet 的 YAML 配置
      tags:
      - K8s StatefulSet 管理
...
swagger: "2.0"